signed/unsigned comparison
git-svn-id: https://svn.code.sf.net/p/nsis/code/NSIS/trunk@5027 212acab6-be3b-0410-9dea-997c60f758d6

+++ Source/Tests/compression.cpp
@@ -19,7 +19,7 @@ public:
 
     for (int i = 0; i < kb; i++) {
       int r = rand();
-      for (int j = 0; j < 1024/sizeof(int); j++) {
+      for (size_t j = 0; j < 1024/sizeof(int); j++) {
         buf.add(&r, sizeof(int));
       }
     }
